+++ backend/src/main/java/com/example/Piroin/project/domain/deposit/service/DepositService.java
@@ -20,7 +20,7 @@ public class DepositService {
     private final AttendanceRepository attendanceRepository;
 
     @Transactional
-    public void recalculateDeposit(Integer userId) {
+    public void recalculateDeposit(Long userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 유저입니다."));
